Struct yaiouom::Measure

pub struct Measure<T, U: Unit> { /* fields omitted */ }

A value with a unit.

Methods

impl<T, U: Unit> Measure<T, U>

pub fn new(value: T) -> Self

Convert from a dimensionless unit.

use yaiouom::*;
use yaiouom::si::*;

let one_meter : Measure<_, Meter> = Measure::new(1);

Warning

This function is somewhat unsafe, as you can use it to build values that are fully polymorphic in unit.

Future versions will make this constructor private and hide it behind syntactic sugar.

pub fn unify<V: Unit>(self) -> Measure<T, V>

Compare two units of measure (not their values).

Out of the box, the Rust type system is not powerful enough to resolve general equality between two units of measure. So, for instance, it will not realize that m * s and s * m are the same unit, or that m / m is the same as the dimensionless unit.

For instance, the following will fail to type:

use yaiouom::*;
use yaiouom::si::*;

let one_meter = Meter::new(1);
let one_second = Second::new(1);

let a = one_meter * one_second;
let b = one_second * one_meter;
assert_eq!(a, b);
// ^^^^^^^^^^^^^^^^^ expected struct `Meter`, found struct `Second`

To work around this, use unify(), as follows:

use yaiouom::*;
use yaiouom::si::*;

let one_meter = Meter::new(1);
let one_second = Second::new(1);

let a = one_meter * one_second;
let b = (one_second * one_meter).unify(); // Delays unit check.
assert_eq!(a, b);

Soundness of unify

If you look at the signature of unify, you can notice that it returns a Measure<T, V> for all V. Despite appearances, this is sound, for two reasons:

A refinement type for units of measure

The companion linter for this crate implements a refinement of the Rust type system dedicated to units of measure. What this means, in practice, is that whenever you call foo.unify(), it will check that the returned type is correct. By opposition to the general Rust type system, it will correctly realize that m * s and s * m are the same, or even that W * m / W is m, even if Wß is a type variable.

Please don't use unify() without the linter :)

Dynamic checks

As a fallback, in debug builds, each call to unify will panic if type V is not equivalent ot type U.

pub fn from<V>(value: Measure<V, U>) -> Self where
T: From<V>,

Convert between two value representations (e.g. u32 vs u64) in the same unit.

Ideally, this should be an implementation of From, but it conflicts with the reflexive implementation.

use yaiouom::*;
use yaiouom::si::*;

let one_meter_usize = Meter::new(1 as i32);
let one_meter_isize : Measure<i64, Meter> = Measure::from(one_meter_usize);

pub fn into<V>(self) -> Measure<V, U> where
T: Into<V>,

Convert between two value representations (e.g. u32 vs u64) In the same unit.

Ideally, this should be an implementation of From, but it conflicts with the reflexive implementation.

use yaiouom::*;
use yaiouom::si::*;

let one_meter_usize : Measure<i32, Meter> = Measure::new(1);
let one_meter_isize : Measure<i64, Meter> = one_meter_usize.into();

pub fn as_runtime(&self) -> RuntimeUnit

Extract a runtime representation of a unit.

This runtime representation is designed mainly for debugging purposes.

use yaiouom::*;
use yaiouom::si::*;

let one_meter_usize : Measure<i32, Meter> = Measure::new(1);
assert_eq!(one_meter_usize.as_runtime().to_string(), "m");

Performance note

This method is fine for debugging, but should not be used in a tight loop.

impl<T> Measure<T, Dimensionless>

pub fn unwrap(self) -> T